Move arg_parser into its own file
